Add config option to disable ipset timeouts.
Ubiquity EdgeRouter 10X does not have any way to create ipsets with
timeouts through the CLI or WEB configuration system.
This makes crowdsec-firewall-bouncer fail to manage the contents of
ipsets created without timeouts.
This commit adds a config option to disable use of timeouts, to make it
possible to use on the EdgeRouter 10X.